+++ src/enhancements.js
@@ -1,3 +1,5 @@
+const srcset = require('srcset');
+
 function imagesAtFullSize(doc) {
 	/*
 		Replace:
@@ -13,10 +15,16 @@ function imagesAtFullSize(doc) {
 		let original = anchor.href;
 
 		// only replace if the HREF matches an image file
-		if (original.match(/\.(png|jpg|jpeg|gif|svg)$/)) {
+		// and exclude wikipedia links to image file pages
+		if (
+			original.match(/\.(png|jpg|jpeg|gif|svg)$/i) &&
+			!original.match(/wiki\/File\:/)
+		) {
 			img.setAttribute('src', original);
 			anchor.parentNode.replaceChild(img, anchor);
 		}
+
+		anchor.classList.add('pcl--figure-like');
 	});
 
 	/*
@@ -41,14 +49,13 @@ function wikipediaSpecific(doc) {
 	).forEach(el => el.remove());
 }
 
+/* 
+	Mark some links as not needing their HREF appended:
+	- links whose text content is the HREF
+	- in-page anchors
+*/
 function noUselessHref(doc) {
-	/* 
-		Mark some links as not needing their HREF appended:
-		- links whose text content is the HREF
-		- in-page anchors
-	*/
-
-	Array.from(doc.querySelectorAll(`a`))
+	Array.from(doc.querySelectorAll('a'))
 		.filter(function(el) {
 			return (
 				(el.getAttribute('href') || '').match(/^\#/) ||
@@ -58,8 +65,44 @@ function noUselessHref(doc) {
 		.forEach(el => el.classList.add('no-href'));
 }
 
+/*
+	Convert relative URIs to absolute URIs:
+	
+	* the `href` attribute of <a> elements (except for in-page anchors)
+	* the `src` attribute of <img> elements
+	* the `srcset` attribute of <source> elements (inside <picture> elements)
+ */
+function relativeToAbsoluteURIs(doc) {
+	function absoluteSrcset(str) {
+		return srcset.stringify(
+			srcset.parse(str).map(item => ({
+				...item,
+				url: new URL(item.url, doc.baseURI).href
+			}))
+		);
+	}
+
+	Array.from(doc.querySelectorAll('a:not([href^="#"])')).forEach(el => {
+		el.setAttribute('href', el.href);
+	});
+
+	Array.from(doc.querySelectorAll('img')).forEach(el => {
+		el.setAttribute('src', el.src);
+	});
+
+	Array.from(doc.querySelectorAll('picture source, img')).forEach(el => {
+		if (el.hasAttribute('srcset')) {
+			el.setAttribute(
+				'srcset',
+				absoluteSrcset(el.getAttribute('srcset'))
+			);
+		}
+	});
+}
+
 module.exports = {
 	imagesAtFullSize,
 	noUselessHref,
-	wikipediaSpecific
+	wikipediaSpecific,
+	relativeToAbsoluteURIs
 };